Utils: atomically write output files
Two generators may be working in the same directory and writing the same
dependent file. There is no guarantee that two programs opening the same
path and writing into it will produce a coherent result.

Fix by writing into a temporary file, then atomically renaming over the
destination, so one complete file always exists.

This has not been tested on Windows, and may not work right if two
generators are intending to write different contents to the file.