Fix typo: jave_exec -> java_exec in microvcodedsgen (#7)
* Fix typo: jave_exec -> java_exec

The variable was defined as `jave_exec` on line 4 but referenced 
as `java_exec` in the conditional check on line 7. This typo would 
cause the script to check an undefined variable, leading to 
incorrect behavior.

Fixed by renaming the variable to `java_exec` throughout.

* Fix typo: jave_exec -> java_exec

The variable was defined as `jave_exec` on line 4 but referenced 
as `java_exec` in the conditional check on line 7. This typo would 
cause the script to check an undefined variable, leading to 
incorrect behavior.

Fixed by renaming the variable to `java_exec` throughout.